requirements.txt
python-telegram-bot==20.3
openai==1.14.3
pandas==2.2.1
gspread==6.1.4
google-auth==2.29.0
python-dotenv==1.0.1

# Platform dependencies
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.36
alembic==1.14.0
bcrypt==4.1.3
itsdangerous==2.2.0
jinja2==3.1.4
python-multipart==0.0.9
aiofiles==23.2.1
pypdf==4.2.0
python-docx==1.1.2
httpx~=0.24.0
